Jakarta Persistence 3.2 - initial commit
