+++ pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2022  Contributors to the Eclipse Foundation
+    Copyright (c) 2022, 2024  Contributors to the Eclipse Foundation
     All rights reserved.
 
     This program and the accompanying materials are made available under the
@@ -93,7 +93,7 @@
         <!-- Jakarta Concurrency -->
         <concurrent-api.version>2.0.0</concurrent-api.version>
         <!-- Jakarta Persistence -->
-        <jakarta-persistence-api.version>3.1.0</jakarta-persistence-api.version>
+        <jakarta-persistence-api.version>3.2.0-M2</jakarta-persistence-api.version>
         <jakarta.activation-api.version>2.0.0</jakarta.activation-api.version>
 
         <jakarta.annotation-api.version>2.1.1</jakarta.annotation-api.version>
